flask-api/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.0.0
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.0.3
flask-bcrypt==0.7.1
flask-migrate==2.2.1
flask-script==2.0.6
flask-sqlalchemy==2.3.2
flask==1.0.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
jinja2==2.10
mako==1.0.7
markupsafe==1.1.1
marshmallow==2.15.3
more-itertools==4.2.0
pluggy==0.7.1
psycopg2==2.7.5
py==1.5.4
pycparser==2.18
pyjwt==1.6.4
pytest-cov==2.5.1
pytest==3.6.4
python-coveralls==2.9.1
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pyyaml==3.13
requests==2.19.1
six==1.11.0
sqlalchemy==1.3.22
urllib3==1.23
werkzeug==0.14.1
